Refactor is_repo_url test
